# refactor: Use sync.Pool to reuse the []byte with specific length