--- a/internal/rdb/structure/int.go
+++ b/internal/rdb/structure/int.go
@@ -3,6 +3,8 @@ package structure
 import (
 	"encoding/binary"
 	"io"
+
+	"RedisShake/internal/log"
 )
 
 func ReadUint8(rd io.Reader) uint8 {
@@ -11,8 +13,12 @@ func ReadUint8(rd io.Reader) uint8 {
 }
 
 func ReadUint16(rd io.Reader) uint16 {
-	buf := ReadBytes(rd, 2)
-	return binary.LittleEndian.Uint16(buf)
+	data := BytesPoolWithCap2.Get().([]byte)
+	defer BytesPoolWithCap2.Put(data)
+	if _, err := io.ReadFull(rd, data); err != nil {
+		log.Panicf(err.Error())
+	}
+	return binary.LittleEndian.Uint16(data)
 }
 
 func ReadUint24(rd io.Reader) uint32 {
@@ -22,13 +28,21 @@ func ReadUint24(rd io.Reader) uint32 {
 }
 
 func ReadUint32(rd io.Reader) uint32 {
-	buf := ReadBytes(rd, 4)
-	return binary.LittleEndian.Uint32(buf)
+	data := BytesPoolWithCap4.Get().([]byte)
+	defer BytesPoolWithCap4.Put(data)
+	if _, err := io.ReadFull(rd, data); err != nil {
+		log.Panicf(err.Error())
+	}
+	return binary.LittleEndian.Uint32(data)
 }
 
 func ReadUint64(rd io.Reader) uint64 {
-	buf := ReadBytes(rd, 8)
-	return binary.LittleEndian.Uint64(buf)
+	data := BytesPoolWithCap8.Get().([]byte)
+	defer BytesPoolWithCap8.Put(data)
+	if _, err := io.ReadFull(rd, data); err != nil {
+		log.Panicf(err.Error())
+	}
+	return binary.LittleEndian.Uint64(data)
 }
 
 func ReadInt8(rd io.Reader) int8 {
@@ -37,8 +51,12 @@ func ReadInt8(rd io.Reader) int8 {
 }
 
 func ReadInt16(rd io.Reader) int16 {
-	buf := ReadBytes(rd, 2)
-	return int16(binary.LittleEndian.Uint16(buf))
+	data := BytesPoolWithCap2.Get().([]byte)
+	defer BytesPoolWithCap2.Put(data)
+	if _, err := io.ReadFull(rd, data); err != nil {
+		log.Panicf(err.Error())
+	}
+	return int16(binary.LittleEndian.Uint16(data))
 }
 
 func ReadInt24(rd io.Reader) int32 {
@@ -48,11 +66,19 @@ func ReadInt24(rd io.Reader) int32 {
 }
 
 func ReadInt32(rd io.Reader) int32 {
-	buf := ReadBytes(rd, 4)
-	return int32(binary.LittleEndian.Uint32(buf))
+	data := BytesPoolWithCap4.Get().([]byte)
+	defer BytesPoolWithCap4.Put(data)
+	if _, err := io.ReadFull(rd, data); err != nil {
+		log.Panicf(err.Error())
+	}
+	return int32(binary.LittleEndian.Uint32(data))
 }
 
 func ReadInt64(rd io.Reader) int64 {
-	buf := ReadBytes(rd, 8)
-	return int64(binary.LittleEndian.Uint64(buf))
+	data := BytesPoolWithCap8.Get().([]byte)
+	defer BytesPoolWithCap8.Put(data)
+	if _, err := io.ReadFull(rd, data); err != nil {
+		log.Panicf(err.Error())
+	}
+	return int64(binary.LittleEndian.Uint64(data))
 }