--- a/internal/rdb/structure/int.go
+++ b/internal/rdb/structure/int.go
@@ -14,11 +14,13 @@ func ReadUint8(rd io.Reader) uint8 {
 
 func ReadUint16(rd io.Reader) uint16 {
 	data := BytesPoolWithCap2.Get().([]byte)
-	defer BytesPoolWithCap2.Put(data)
+	BytesPoolWithCap2.Put(data)
 	if _, err := io.ReadFull(rd, data); err != nil {
 		log.Panicf(err.Error())
 	}
-	return binary.LittleEndian.Uint16(data)
+	result := binary.LittleEndian.Uint16(data)
+	BytesPoolWithCap2.Put(data)
+	return result
 }
 
 func ReadUint24(rd io.Reader) uint32 {
@@ -29,20 +31,22 @@ func ReadUint24(rd io.Reader) uint32 {
 
 func ReadUint32(rd io.Reader) uint32 {
 	data := BytesPoolWithCap4.Get().([]byte)
-	defer BytesPoolWithCap4.Put(data)
 	if _, err := io.ReadFull(rd, data); err != nil {
 		log.Panicf(err.Error())
 	}
-	return binary.LittleEndian.Uint32(data)
+	result := binary.LittleEndian.Uint32(data)
+	BytesPoolWithCap4.Put(data)
+	return result
 }
 
 func ReadUint64(rd io.Reader) uint64 {
 	data := BytesPoolWithCap8.Get().([]byte)
-	defer BytesPoolWithCap8.Put(data)
 	if _, err := io.ReadFull(rd, data); err != nil {
 		log.Panicf(err.Error())
 	}
-	return binary.LittleEndian.Uint64(data)
+	result := binary.LittleEndian.Uint64(data)
+	BytesPoolWithCap8.Put(data)
+	return result
 }
 
 func ReadInt8(rd io.Reader) int8 {
@@ -52,11 +56,12 @@ func ReadInt8(rd io.Reader) int8 {
 
 func ReadInt16(rd io.Reader) int16 {
 	data := BytesPoolWithCap2.Get().([]byte)
-	defer BytesPoolWithCap2.Put(data)
 	if _, err := io.ReadFull(rd, data); err != nil {
 		log.Panicf(err.Error())
 	}
-	return int16(binary.LittleEndian.Uint16(data))
+	result := int16(binary.LittleEndian.Uint16(data))
+	BytesPoolWithCap2.Put(data)
+	return result
 }
 
 func ReadInt24(rd io.Reader) int32 {
@@ -67,18 +72,20 @@ func ReadInt24(rd io.Reader) int32 {
 
 func ReadInt32(rd io.Reader) int32 {
 	data := BytesPoolWithCap4.Get().([]byte)
-	defer BytesPoolWithCap4.Put(data)
 	if _, err := io.ReadFull(rd, data); err != nil {
 		log.Panicf(err.Error())
 	}
-	return int32(binary.LittleEndian.Uint32(data))
+	result := int32(binary.LittleEndian.Uint32(data))
+	BytesPoolWithCap4.Put(data)
+	return result
 }
 
 func ReadInt64(rd io.Reader) int64 {
 	data := BytesPoolWithCap8.Get().([]byte)
-	defer BytesPoolWithCap8.Put(data)
 	if _, err := io.ReadFull(rd, data); err != nil {
 		log.Panicf(err.Error())
 	}
-	return int64(binary.LittleEndian.Uint64(data))
+	result := int64(binary.LittleEndian.Uint64(data))
+	BytesPoolWithCap8.Put(data)
+	return result
 }