--- a/internal/rdb/structure/int.go
+++ b/internal/rdb/structure/int.go
@@ -13,12 +13,11 @@ func ReadUint8(rd io.Reader) uint8 {
 }
 
 func ReadUint16(rd io.Reader) uint16 {
-	data := BytesPoolWithCap2.Get().([]byte)
-	BytesPoolWithCap2.Put(data)
-	if _, err := io.ReadFull(rd, data); err != nil {
+	data := BytesPoolWithCap2.Get().(*[]byte)
+	if _, err := io.ReadFull(rd, *data); err != nil {
 		log.Panicf(err.Error())
 	}
-	result := binary.LittleEndian.Uint16(data)
+	result := binary.LittleEndian.Uint16(*data)
 	BytesPoolWithCap2.Put(data)
 	return result
 }
@@ -30,21 +29,21 @@ func ReadUint24(rd io.Reader) uint32 {
 }
 
 func ReadUint32(rd io.Reader) uint32 {
-	data := BytesPoolWithCap4.Get().([]byte)
-	if _, err := io.ReadFull(rd, data); err != nil {
+	data := BytesPoolWithCap4.Get().(*[]byte)
+	if _, err := io.ReadFull(rd, *data); err != nil {
 		log.Panicf(err.Error())
 	}
-	result := binary.LittleEndian.Uint32(data)
+	result := binary.LittleEndian.Uint32(*data)
 	BytesPoolWithCap4.Put(data)
 	return result
 }
 
 func ReadUint64(rd io.Reader) uint64 {
-	data := BytesPoolWithCap8.Get().([]byte)
-	if _, err := io.ReadFull(rd, data); err != nil {
+	data := BytesPoolWithCap8.Get().(*[]byte)
+	if _, err := io.ReadFull(rd, *data); err != nil {
 		log.Panicf(err.Error())
 	}
-	result := binary.LittleEndian.Uint64(data)
+	result := binary.LittleEndian.Uint64(*data)
 	BytesPoolWithCap8.Put(data)
 	return result
 }
@@ -55,11 +54,11 @@ func ReadInt8(rd io.Reader) int8 {
 }
 
 func ReadInt16(rd io.Reader) int16 {
-	data := BytesPoolWithCap2.Get().([]byte)
-	if _, err := io.ReadFull(rd, data); err != nil {
+	data := BytesPoolWithCap2.Get().(*[]byte)
+	if _, err := io.ReadFull(rd, *data); err != nil {
 		log.Panicf(err.Error())
 	}
-	result := int16(binary.LittleEndian.Uint16(data))
+	result := int16(binary.LittleEndian.Uint16(*data))
 	BytesPoolWithCap2.Put(data)
 	return result
 }
@@ -71,21 +70,21 @@ func ReadInt24(rd io.Reader) int32 {
 }
 
 func ReadInt32(rd io.Reader) int32 {
-	data := BytesPoolWithCap4.Get().([]byte)
-	if _, err := io.ReadFull(rd, data); err != nil {
+	data := BytesPoolWithCap4.Get().(*[]byte)
+	if _, err := io.ReadFull(rd, *data); err != nil {
 		log.Panicf(err.Error())
 	}
-	result := int32(binary.LittleEndian.Uint32(data))
+	result := int32(binary.LittleEndian.Uint32(*data))
 	BytesPoolWithCap4.Put(data)
 	return result
 }
 
 func ReadInt64(rd io.Reader) int64 {
-	data := BytesPoolWithCap8.Get().([]byte)
-	if _, err := io.ReadFull(rd, data); err != nil {
+	data := BytesPoolWithCap8.Get().(*[]byte)
+	if _, err := io.ReadFull(rd, *data); err != nil {
 		log.Panicf(err.Error())
 	}
-	result := int64(binary.LittleEndian.Uint64(data))
+	result := int64(binary.LittleEndian.Uint64(*data))
 	BytesPoolWithCap8.Put(data)
 	return result
 }