--- a/internal/rdb/structure/ziplist.go
+++ b/internal/rdb/structure/ziplist.go
@@ -80,11 +80,11 @@ func readZipListEntry(rd io.Reader, firstByte byte) string {
 		length := (int(firstByte&0x3f) << 8) | int(secondByte)
 		return string(ReadBytes(rd, length))
 	case zipStr32B:
-		lenBytes := BytesPoolWithCap4.Get().([]byte)
-		if _, err := io.ReadFull(rd, lenBytes); err != nil {
+		lenBytes := BytesPoolWithCap4.Get().(*[]byte)
+		if _, err := io.ReadFull(rd, *lenBytes); err != nil {
 			log.Panicf(err.Error())
 		}
-		length := binary.BigEndian.Uint32(lenBytes)
+		length := binary.BigEndian.Uint32(*lenBytes)
 		BytesPoolWithCap4.Put(lenBytes)
 		return string(ReadBytes(rd, int(length)))
 	}